+++ ext/Server/Bot/VehicleJetControl.lua
@@ -39,7 +39,7 @@ function VehicleJetControl:UpdateMovementJet(p_DeltaTime, p_Bot)
 		end
 	end
 
-	p_Bot._JetMidPosition = g_GameDirector._MidMapPoint:Clone()
+	p_Bot._JetMidPosition = g_GameDirector:GetActiveTargetPointPosition(p_Bot.m_Player.teamId)
 	p_Bot._JetMidPosition.y = p_Bot._JetMidPosition.y + Registry.VEHICLES.JET_TARGET_HEIGHT
 	if (p_Bot.m_Player.teamId % 2) == 1 then
 		p_Bot._JetMidPosition.z = p_Bot._JetMidPosition.z + 100