+++ .github/CHANGELOG.md
@@ -19,19 +19,18 @@ This is the changelog for the version V3.0 Don't forget to [join us on Discord](
 * Separate worsening values for different vehicle types (by MatiasPastori)
 * Make bots attack the gunship by themselves (by Tagener Noisu)
 * Bots can use every jet in conquest now (no more paths needed)
+* No more air-paths needed (only enter-paths for choppers)
 
 ### some open TODOs:
 * rework of raycasts for better performance
 * fully support default-sapwn-method (for now only on TDM/GM/SDM by default)
 * further performance increses
-* jet navigation and driving improvements (get rid of paths already possible!)
-* get Rid of chopper-paths if possible -> also handle different hights and positions depending on caturep-points
+* chopper: handle different hights and positions depending on capture-points
 * performance of spawn-point parsing -> save those?
 * performance of node-editor
 * add logic for jets in rush
 * improve team-logic for AA and other vehicles (example AA on Rush)
 * improve node-editor
-* Choppers: Abort attack if too steep or low
 
 ### Bug fixes
 * fix stationary AA again...